Create README.md
Currently im an university student on Chile, my dream are going out to experience working with people around the world, also working with AI to built a better society for the next 10 years and so on. If this dream fails i dont have any fear, i have also other dreams that will keep going to be here in the world of Computer Science. Love to all <3 :)